0.1.2e
Started to zero in on the issue, but still don't have an answer.
When MathHelper.DrawCube() is trying to check the neighboring blocks to check if this particular block has visible faces, it is using "position" this is passed to it by the Block.Draw() which takes position as the local variable in that instance of Block, which itself is set by Block.SetPosition() called in the chunk class when the chunk is creating the blocks in its Start() function. The problem is that when MathHelper accesses "position" it is a totally different value than what it should be, but Chunk.pos in Chunk.Start() is the correct value, I tested it, so between when it is assigned to the block instance via SetPos() and accessed by MathHelper.DrawBlock() that is where things are getting fishy. idk not sure