0.1.2e
Somewhere between when the Position of each block instance is set in Chunk.Start() -> Block.SetPosition() and when it is accessed in MathHelper.DrawCube() -> Int3 position = block.Position, there is a problem. The position it is showing is completely wrong. But when I check the value of pos in Chunk.Start() as it is generating the Blocks and then using pos to assign it to the Blocks, it seems to be using the correct value at that point. Not sure where the problem is happening.
The Log shows this:
Chunk HAS ALL neighbors: C_-1_-1_-1
Neighbors Checked: (X:-2, Y:-1, Z:-1), (X:0, Y:-1, Z:-1), (X:-1, Y:-2, Z:-1), (X:-1, Y:0, Z:-1), (X:-1, Y:-1, Z:-2), (X:-1, Y:-1, Z:0)
Checking Faces of Block: 0, 0, 0
At World Pos: 31, 3, 29

0,0,0 of C_-1_-1_-1 should be -16,-16,-16 not 31, 3, 29.....